slugutil: change bands texture format from RGBA32UI to RG16UI (4x smaller)

+++ libs/slugutil/slugutil.h
@@ -9,7 +9,7 @@
 
 typedef struct {
     vec2_t p[3];
-    uint32_t texture[2];
+    uint16_t texture[2];
 } slug_curve_t;
 
 typedef struct {
